src/main/java/ua/bykadoroff/hw13/BubbleSort.java
package ua.bykadoroff.hw13;

public class BubbleSort {

    public static int[] bubbleSort(int[] arr) {
        int[] sortedArray = arr.clone();

        for (int i = 0; i < sortedArray.length; i++) {
            for (int j = 1; j < sortedArray.length - i; j++) {
                if (sortedArray[j] < sortedArray[j - 1]) {
                    int temp = sortedArray[j - 1];
                    sortedArray[j - 1] = sortedArray[j];
                    sortedArray[j] = temp;
                }
            }
        }
        return sortedArray;
    }
}
